+++ migration/statedifflength/migrator.go
@@ -22,8 +22,8 @@ import (
 
 const (
 	maxConcurrency       = 8
-	defaultBatchByteSize = 24 * db.Megabyte
-	targetBatchByteSize  = 16 * db.Megabyte
+	defaultBatchByteSize = 128 * db.Megabyte
+	targetBatchByteSize  = 96 * db.Megabyte
 )
 
 var (
@@ -91,11 +91,19 @@ func (m *Migrator) Migrate(
 
 	source := pipeline.Source(blockRange(start, height))
 	readers := pipeline.New(
-		source, concurrency, newIngestor(database, batchSemaphore,
-			concurrency, min(batchByteSize, targetBatchByteSize)),
+		source,
+		concurrency,
+		newIngestor(
+			database,
+			batchSemaphore,
+			concurrency,
+			min(batchByteSize, targetBatchByteSize),
+		),
 	)
 	committed := pipeline.New(
-		readers, 1, newCommitter(logger, batchSemaphore, height),
+		readers,
+		1,
+		newCommitter(logger, batchSemaphore, height),
 	)
 
 	_, wait := committed.Run(ctx)
@@ -133,23 +141,4 @@ func blockRange(start, end uint64) iter.Seq[uint64] {
 			}
 		}
 	}
-}
-
-// backfillBlock derives the state diff length from the block's state update and
-// writes the block's commitments with it into batch. A missing commitments or
-// state update record above the pruned prefix means the two buckets disagree, i.e.
-// the database is corrupt, so the read error is returned rather than skipped.
-func backfillBlock(r db.KeyValueReader, batch db.KeyValueWriter, blockNumber uint64) error {
-	commitments, err := core.GetBlockCommitmentByBlockNum(r, blockNumber)
-	if err != nil {
-		return fmt.Errorf("getting commitments for block %d: %w", blockNumber, err)
-	}
-
-	stateUpdate, err := core.GetStateUpdateByBlockNum(r, blockNumber)
-	if err != nil {
-		return fmt.Errorf("getting state update for block %d: %w", blockNumber, err)
-	}
-
-	commitments.StateDiffLength = stateUpdate.StateDiff.Length()
-	return core.WriteBlockCommitment(batch, blockNumber, commitments)
 }
